Make base_pc 64-bit and use full PC for RV64 jump/branch instructions
Changes:
- Changed base_pc from u32 to u64 to support full 64-bit addresses
- Updated AUIPC: Uses I64Const with full 64-bit PC when RV64 enabled
- Updated JAL: Stores full 64-bit return address when RV64 enabled
- Updated JALR: Stores full 64-bit return address and uses 64-bit mask when RV64 enabled
- Updated branch instructions: Uses I64 comparison instructions when RV64 enabled
- Updated pc_to_func_idx and jump_to_pc to accept u64 PC values
- Fixed all test files to use u64 for base_pc parameter

All 45 tests pass (35 unit + 10 integration).

+++ crates/speet-riscv/tests/hint_tracking_test.rs
@@ -50,7 +50,7 @@ fn test_hint_tracking_with_rv32im_multiply() {
             ty: TypeIdx(0),
         },
         None,
-        start_addr as u32,
+        start_addr as u64,
         true, // Enable HINT tracking
         false, // Disable RV64
         false, // Disable memory64
